% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{addCatchDuringCDColumn}
\alias{addCatchDuringCDColumn}
\title{addCatchDuringCDColumn}
\usage{
addCatchDuringCDColumn(data, pokedex)
}
\arguments{
\item{data}{history of catches}

\item{pokedex}{pokedex}
}
\description{
addCatchDuringCDColumn
}
