% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{addShinyColumns}
\alias{addShinyColumns}
\title{addShinyColumns}
\usage{
addShinyColumns(data, pokedex)
}
\arguments{
\item{data}{history of catches}

\item{pokedex}{pokedex}
}
\description{
addShinyColumns
}
