% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{countCapturedByDay}
\alias{countCapturedByDay}
\title{countCapturedByDay}
\usage{
countCapturedByDay(data, day)
}
\arguments{
\item{data}{history of catch}

\item{day}{Day}
}
\description{
countCapturedByDay
}
