% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{countCapturedPokemonByDay}
\alias{countCapturedPokemonByDay}
\title{countCapturedPokemonByDay}
\usage{
countCapturedPokemonByDay(data, day, pokemonNr)
}
\arguments{
\item{data}{history of catch}

\item{day}{Day}

\item{pokemonNr}{Pokemon Nr}
}
\description{
countCapturedPokemonByDay
}
