% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{countFoundShinyByDay}
\alias{countFoundShinyByDay}
\title{countFoundShinyByDay}
\usage{
countFoundShinyByDay(data, day)
}
\arguments{
\item{data}{history of catch}

\item{day}{day to count shinys}
}
\description{
countFoundShinyByDay
}
