% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{countShinyByDay}
\alias{countShinyByDay}
\title{countShinyByDay}
\usage{
countShinyByDay(data, day)
}
\arguments{
\item{data}{history of catch}

\item{day}{day to count shinys}
}
\description{
countShinyByDay
}
