% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{EstimaProporExact}
\alias{EstimaProporExact}
\title{EstimaProporExact}
\usage{
EstimaProporExact(k, n, alpha)
}
\arguments{
\item{k}{nombre reussite}

\item{n}{taille echantillon}

\item{alpha}{seuil}
}
\description{
EstimaProporExact
}
