% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getCaptureByDay}
\alias{getCaptureByDay}
\title{getCaptureByDay}
\usage{
getCaptureByDay(day, data)
}
\arguments{
\item{day}{Day 2019-12-31}

\item{data}{All history of pokemon}
}
\description{
getCaptureByDay
}
