% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getPokemonCatchDuringCD}
\alias{getPokemonCatchDuringCD}
\title{getPokemonCatchDuringCD}
\usage{
getPokemonCatchDuringCD(data)
}
\arguments{
\item{data}{of catch}
}
\description{
getPokemonCatchDuringCD
}
