% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getStatBetweenDates}
\alias{getStatBetweenDates}
\title{getTotalShinyProbaDuringCD}
\usage{
getStatBetweenDates(data, date1, date2)
}
\arguments{
\item{data}{history of catch}
}
\description{
getTotalShinyProbaDuringCD
}
