% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getTotalShinyProbaDuringCD}
\alias{getTotalShinyProbaDuringCD}
\title{getTotalShinyProbaDuringCD}
\usage{
getTotalShinyProbaDuringCD(data)
}
\arguments{
\item{data}{history of catch}
}
\description{
getTotalShinyProbaDuringCD
}
