% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getTotalShinyProbaOutCD}
\alias{getTotalShinyProbaOutCD}
\title{getTotalShinyProbaOutCD}
\usage{
getTotalShinyProbaOutCD(data)
}
\arguments{
\item{data}{history of catch}
}
\description{
getTotalShinyProbaOutCD
}
