% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{installGitHubPackage}
\alias{installGitHubPackage}
\title{installandLoadGitHubPackage}
\usage{
installGitHubPackage(user_repo_pat)
}
\arguments{
\item{user_repo_pat}{user/repo/pat}
}
\description{
installandLoadGitHubPackage
}
