% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{installRPackages}
\alias{installRPackages}
\title{installPackages}
\usage{
installRPackages(packages_to_install)
}
\arguments{
\item{packages_to_install}{list of R packages to install}
}
\description{
installPackages
}
