% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{libraryGitHubPackage}
\alias{libraryGitHubPackage}
\title{libraryGitHubPackage}
\usage{
libraryGitHubPackage(user_repo_pat)
}
\arguments{
\item{user_repo_pat}{user/repo/pat if private}
}
\description{
libraryGitHubPackage
}
