% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{NshinyOdd}
\alias{NshinyOdd}
\title{NshinyOdd}
\usage{
NshinyOdd(nPotentialShinyCatched, nshiny)
}
\arguments{
\item{nPotentialShinyCatched}{number of shiny captured}

\item{nshiny}{number of shiny catched}
}
\description{
NshinyOdd
}
