% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plotCapturePokemonByDay}
\alias{plotCapturePokemonByDay}
\title{plotCapturePokemonByDay}
\usage{
plotCapturePokemonByDay(data, pokemonNr, pokedex)
}
\arguments{
\item{data}{data to plot from}

\item{pokemonNr}{nr of the pokemon to plot}

\item{pokedex}{pokedex}
}
\description{
plotCapturePokemonByDay
}
