% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{setIsShinyCD}
\alias{setIsShinyCD}
\title{setIsShinyCD}
\usage{
setIsShinyCD(rowNumber, data, value)
}
\arguments{
\item{rowNumber}{row to change}

\item{data}{history of capture}

\item{value}{value to put}
}
\description{
setIsShinyCD
}
