% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{summaryPokemonByDay}
\alias{summaryPokemonByDay}
\title{summaryPokemonByDay}
\usage{
summaryPokemonByDay(data, pokemonNr)
}
\arguments{
\item{data}{history of catch}

\item{pokemonNr}{number of pokemon}
}
\description{
summaryPokemonByDay
}
